Fix gem spec and install task.
It is gauche to invoke `sudo` in a Rakefile; @alrs told me about this about
a decade ago, but I appear not to have actually made the change.

Gem::Specification#has_rdoc is deprecated, so that has been removed.

Fun story:  if you don't clean out the package directory, apparently
the package task will tell you something like this:
  Gem::InvalidSpecificationException: ["doc/examples/return.ru"] are not files
Having been fairly certain it *was* a file, this was mystifying.  So I walked
down the stack, but did not have to walk far, as the source of the error was
first:  rubygems/specification_policy.rb:298.  Using the Kernighan methodⁱ, I
noted that File.file?("doc/examples/return.ru") was returning false.  I checked
that the method hadn't been monkey-patched, I checked the contents of the
directory, and the file wasn't there.  I was squinting at the screen:  I
double-checked the Rakefile, and it added every existing file under doc/ to
the list of files.  Just a few minutes before, I had managed to get Ruby to
tell me `syntax error, unexpected '}\, expecting '}'` in response
to a typo, so I was starting to have some grave doubts about the ground under
my feet.  At any rate, I am not certain why Gem::PackageTask does not pick up
new files but sometimes if the computer tells you a file does not exist, it
is correct.