Rack::Lint stopped accepting canonical header case.
Some time in the last few years, Rack::Lint started requiring header
names to be lower-case rather than the canonical case used by every
web server and user agent, the same case style used in the RFCs, all
documentation everywhere, since the 1990s.  For the mandatory lower-
case in HTTP/2, they still have to convert it to be compliant, but
that is for the sake of the compression algorithm (imagine sending
so many headers that you feel the need to compress them), which is
to say that forcing lower-case is for machines, not people:  HTTP/2
isn't even a plain-text protocol, so forcing the wire format into
the application programmer's lap is obtuse.